+++ Personal_Finance_Management_System/src/main/java/User_Authentication/User.java
@@ -1,12 +1,12 @@
 package User_Authentication;
 
-import static User_Authentication.DataBase.Connection;
+
 import java.sql.Connection;
 import java.sql.DriverManager;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
-import java.sql.Statement;
+
 import java.time.LocalDate;
 import java.util.ArrayList;
